Restructure routes of balances
The new structure introduces an "index" route,
which allows us to place reuseable items such
as themes at the top-level of this route
section. We also got rid of the Cards and use
our new Title and SubTitle component for styling.
