--- src/routes/App/routes/balances/components/BalanceOverview/BalanceCard/CardTheme.scss
@@ -1,4 +1,0 @@
-.card {
-  margin: 1rem;
-  width: auto !important;
-}
