--- src/routes/App/routes/balances/components/BalanceOverview/BalanceCard/CardTitleTheme.scss
@@ -1,15 +1,0 @@
-.title {
-
-  overflow: hidden;
-  text-overflow: ellipsis;
-  margin-bottom: 0 !important;
-}
-
-.cardTitle {
-  > div {
-    overflow: hidden;
-  }
-
-  padding-top: 1rem !important;
-  padding-bottom: 0 !important;
-}
