--- src/routes/App/routes/balances/components/BalanceOverview/BalanceCard/TransactionListItemTheme.scss
@@ -1,14 +1,0 @@
-@import "src/styles/personListItem";
-
-// Using the "legend" property of the ListItem applies a fixed height on the itemContentRoot.
-.itemContentRoot {
-  height: auto !important;
-}
-
-.listItem:first-of-type {
-  margin-top: 1rem;
-}
-
-.listItem:last-of-type {
-  margin-bottom: 1rem;
-}
