--- src/routes/App/routes/balances/components/BalanceOverview/BalanceCard/TransactionListTheme.scss
@@ -1,3 +1,0 @@
-.list {
-  padding: 0;
-}
