--- src/routes/App/routes/balances/components/BalanceOverview/BalanceOverview.scss
@@ -1,47 +1,0 @@
-.balanceContainer {
-
-  display: flex;
-  justify-content: space-between;
-
-  font-size: larger;
-
-  margin: 1rem;
-  padding: 1rem;
-
-  background-color: white;
-  border-radius: .2rem;
-
-  .debtor {
-    flex-grow: 1;
-  }
-
-  .sign {
-    display: flex;
-    align-items: center;
-  }
-
-  .amount {
-    margin-left: 1rem;
-  }
-
-  animation: FadeIn 0.6s cubic-bezier(0.45, 0.05, 0.55, 0.95);
-  animation-fill-mode: both;
-}
-
-@keyframes FadeIn {
-  0% {
-    opacity: 0;
-    transform: translateX(-100%);
-  }
-
-  100% {
-    opacity: 1;
-    transform: translateX(0);
-  }
-}
-
-@for $i from 1 to 100 {
-  .balanceList {
-    .balanceContainer:nth-child(#{$i}) { animation-delay: $i * 0.05s; }
-  }
-}
