+++ src/routes/App/routes/balances/index.js
@@ -1,14 +1,11 @@
-import BalanceOverview from "./components/BalanceOverview";
-import NameRoute from "./routes/name"
-import withNavigation from "../../../../components/hoc/withNavigation";
+import IndexRoute from "./routes/index"
+import NameRoute from "./routes/name";
 import EmptyLayout from "../../../../layouts/EmptyLayout";
 
 export default {
   path: "balances",
   component: EmptyLayout,
-  indexRoute: {
-    component: withNavigation(BalanceOverview)
-  },
+  indexRoute: IndexRoute,
   childRoutes: [
     NameRoute
   ]
