+++ src/routes/App/routes/balances/routes/index/index.js
@@ -1,0 +1,6 @@
+import BalanceOverview from "./components/BalanceOverview/BalanceOverview";
+import withNavigation from "../../../../../../components/hoc/withNavigation";
+
+export default {
+  component: withNavigation(BalanceOverview)
+}
