--- src/routes/App/routes/balances/routes/name/components/PersonBalance/AmountTheme.scss
@@ -1,7 +1,0 @@
-.positive {
-  color: green;
-}
-
-.negative {
-  color: red;
-}
