--- src/routes/App/routes/balances/routes/name/components/PersonBalance/AppBarTheme.scss
@@ -1,8 +1,0 @@
-.appBar {
-  color: white;
-}
-
-.title {
-  text-align: center;
-  text-transform: uppercase;
-}
