--- src/routes/App/routes/balances/routes/name/components/PersonBalance/InfoMessageCardTextTheme.scss
@@ -1,3 +1,0 @@
-.cardText.cardText {
-  padding-bottom: 0;
-}
