--- src/routes/App/routes/balances/routes/name/components/PersonBalance/InfoMessageCardTitleTheme.scss
@@ -1,3 +1,0 @@
-.cardTitle.cardTitle {
-  padding-bottom: 0;
-}
