+++ src/routes/App/routes/balances/routes/name/components/PersonBalance/PersonBalance.js
@@ -13,18 +13,15 @@ import {
   makeGetPaidBills,
   makeGetRelatedBills
 } from "../../../../../../modules/currentNobt/selectors";
-import AmountTheme from "./AmountTheme.scss";
-import AppBarTheme from "./AppBarTheme.scss";
 import LocationBuilder from "../../../../../../modules/navigation/LocationBuilder";
 import { FontIcon } from "react-toolbox/lib/font_icon/index";
 import HeadRoom from "react-headroom";
 import InfoMessageListTheme from "./InfoMessageListTheme.scss";
-import InfoMessageCardTextTheme from "./InfoMessageCardTextTheme.scss";
-import InfoMessageCardTitleTheme from "./InfoMessageCardTitleTheme.scss";
 import AsyncActionStatus from "../../../../../../../../const/AsyncActionStatus";
-import { Card, CardText, CardTitle } from "react-toolbox/lib/card";
 import { FormattedMessage } from "react-intl";
-
+import { SubTitle, Title } from "components/text";
+import AppBarTheme from "../../../../themes/AppBarTheme.scss"
+import AmountTheme from "../../../../themes/AmountTheme.scss"
 
 class PersonBalance extends React.Component {
 
@@ -44,94 +41,84 @@ class PersonBalance extends React.Component {
             />
           </HeadRoom>
 
+          <div>
+            <Title>Summary</Title>
 
-          <Card>
-            <CardTitle title="Summary" theme={InfoMessageCardTitleTheme} />
-            <CardText theme={InfoMessageCardTextTheme}>
-              <List theme={InfoMessageListTheme}>
-                <ListItem
-                  ripple={false}
-                  leftIcon="info_outline"
-                  itemContent={(<FormattedMessage
-                      id="PersonBalance.paidBillsSummary"
-                      defaultMessage="{name} {numberOfBills, plural,
+            <List theme={InfoMessageListTheme}>
+              <ListItem
+                ripple={false}
+                leftIcon="info_outline"
+                itemContent={(<FormattedMessage
+                    id="PersonBalance.paidBillsSummary"
+                    defaultMessage="{name} {numberOfBills, plural,
                                             =0 {did not pay any bills}
                                             =1 {paid 1 bill ({totalAmount})}
                                             other {paid {numberOfBills} bills ({totalAmount})}}."
-                      values={{
-                        name: this.props.balance.me.name,
-                        numberOfBills: this.props.paidBills.length,
-                        totalAmount: <Amount value={this.props.sumOfPaidBills} absolute={true} />
-                      }}
-                    />
-                  )}
-                />
-
-                <ListItem
-                  ripple={false}
-                  leftIcon="info_outline"
-                  itemContent={
-                    <FormattedMessage
-                      id="PersonBalance.paidBillsSummary"
-                      defaultMessage={
-                        this.props.bills.length !== this.props.relatedBills.length ?
-                          "{name} participates in {numberOfBills} of {allBills} bills." :
-                          "{name} participates in all {allBills} bills."
-                      }
-                      values={{
-                        name: this.props.balance.me.name,
-                        allBills: this.props.bills.length,
-                        numberOfBills: this.props.relatedBills.length
-                      }}
-                    />
-                  }
-                />
-
-                <ListItem
-                  ripple={false}
-                  leftIcon="info_outline"
-                  itemContent={(<FormattedMessage
-                      id="PersonBalance.summary"
-                      defaultMessage="{name} {verb} {amount} {preposition} {numberOfOtherPeople} {numberOfOtherPeople, plural,
-                                                                                                =1 {person}
-                                                                                                other {persons}}."
-                      values={{
-                        name: this.props.balance.me.name,
-                        verb: this.props.balance.me.amount > 0 ? "gets" : "owes",
-                        amount: <Amount value={this.props.balance.me.amount} absolute={true} />,
-                        preposition: this.props.balance.me.amount > 0 ? "from" : "to",
-                        numberOfOtherPeople: this.props.balance.persons.length
-                      }}
-                    />
-                  )}
-                />
-              </List>
-            </CardText>
-          </Card>
-
-          <Card>
-            <CardTitle
-              theme={InfoMessageCardTitleTheme}
-              title="Transactions"
-              // subtitle="The following transactions need to be carried out in order to settle all debts."
-            />
-            <CardText theme={InfoMessageCardTextTheme}>
-              <HOList
-                items={this.props.balance.persons}
-                renderItem={ (transaction) => (
-                  <ListItem
-                    leftActions={[
-                      <Avatar name={transaction.name} medium />
-                    ]}
-                    ripple={false}
-                    key={transaction.name}
-                    caption={transaction.name}
-                    legend={<Amount theme={AmountTheme} value={transaction.amount} absolute={false} />}
+                    values={{
+                      name: this.props.balance.me.name,
+                      numberOfBills: this.props.paidBills.length,
+                      totalAmount: <Amount value={this.props.sumOfPaidBills} absolute={true} />
+                    }}
                   />
                 )}
               />
-            </CardText>
-          </Card>
+
+              <ListItem
+                ripple={false}
+                leftIcon="info_outline"
+                itemContent={
+                  <FormattedMessage
+                    id="PersonBalance.paidBillsSummary"
+                    defaultMessage={
+                      this.props.bills.length !== this.props.relatedBills.length ?
+                        "{name} participates in {numberOfBills} of {allBills} bills." :
+                        "{name} participates in all {allBills} bills."
+                    }
+                    values={{
+                      name: this.props.balance.me.name,
+                      allBills: this.props.bills.length,
+                      numberOfBills: this.props.relatedBills.length
+                    }}
+                  />
+                }
+              />
+            </List>
+          </div>
+
+          <div>
+            <Title>Transactions</Title>
+            <SubTitle>
+              <FormattedMessage
+                id="PersonBalance.summary"
+                defaultMessage="{name} {verb} {amount} {preposition} {numberOfOtherPeople} {numberOfOtherPeople, plural,
+                                                                                                =1 {person}
+                                                                                                other {persons}}."
+                values={{
+                  name: this.props.balance.me.name,
+                  verb: this.props.balance.me.amount > 0 ? "gets" : "owes",
+                  amount: <Amount value={this.props.balance.me.amount} absolute={true} />,
+                  preposition: this.props.balance.me.amount > 0 ? "from" : "to",
+                  numberOfOtherPeople: this.props.balance.persons.length
+                }}
+              />
+            </SubTitle>
+
+            <HOList
+              items={this.props.balance.persons}
+              renderItem={ (transaction) => (
+                <ListItem
+                  leftActions={[
+                    <Avatar name={transaction.name} medium />
+                  ]}
+                  ripple={false}
+                  key={transaction.name}
+                  caption={transaction.name}
+                  legend={<Amount theme={AmountTheme} value={transaction.amount} absolute={false} />}
+                />
+              )}
+            />
+          </div>
+
         </div>
       )}
