+++ src/routes/index.js
@@ -1,5 +1,5 @@
 // We only need to import the modules necessary for initial render
-import CoreLayout from "../layouts/CoreLayout/CoreLayout";
+import CoreLayout from "../layouts/CoreLayout";
 import LandingRouteFactory from "./Landing";
 import AppRouteFactory from "./App";
 import CreateNobtRouteFactory from "./CreateNobt"
